# swaggerDocument.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "SPCA API"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
    "/users": {
      "get": {
        "tags": ["User"],
        "summary": "Retrieves a list of all users",
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["User"],
        "summary": "Creates a new user",
        "parameters": [
          {
            "name": "centreId",
            "in": "query",
            "description": "ID of the centre to which the user belongs",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "name": "userType",
            "in": "query",
            "description": "Type of user (e.g. Volunteer, Vet)",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "User object that needs to be added",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uri",
                  "example": "https://localhost:8080/centres/1/users/1"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "409": {
            "description": "An existing user already exists with the same username"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["User"],
        "summary": "Logs in a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The user's username",
                    "example": "Steve123"
                  },
                  "password": {
                    "type": "string",
                    "description": "The user's password",
                    "example": "BiggestSimpInAvengers"
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "An access token for the authenticated user"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "description": "The user's ID"
                        },
                        "username": {
                          "type": "string",
                          "description": "The user's username"
                        },
                        "role": {
                          "type": "string",
                          "description": "The user's role (e.g. 'Admin', 'Vet', 'Volunteer')"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "The user's creation date"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "The user's last update date"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users/{userId}": {
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "description": "ID of the user to retrieve or delete",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": ["User"],
        "summary": "Retrieves a single user by ID",
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": ["User"],
        "summary": "Updates user details",
        "description": "Updates the details of a user with the specified ID",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User details updated successfully"
          },
          "400": {
            "description": "Invalid request body"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Deletes a user by ID",
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/centres/{centreId}/scales": {
      "parameters": [
        {
          "name": "centreId",
          "in": "path",
          "description": "ID of the centre to retrieve scales from",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": ["Scale"],
        "summary": "Retrieves a list of all scales in a specific centre",
        "responses": {
          "200": {
            "description": "A list of scales",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Scale"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Centre not found"
          }
        }
      }
    },
    "/centres/{centreId}/scales/{scaleId}": {
      "parameters": [
        {
          "name": "centreId",
          "in": "path",
          "description": "ID of the centre the scale belongs to",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "scaleId",
          "in": "path",
          "description": "ID of the scale to retrieve",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": ["Scale"],
        "summary": "Retrieves a single scale by ID within a specific centre",
        "responses": {
          "200": {
            "description": "Scale found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scale"
                }
              }
            }
          },
          "404": {
            "description": "Scale / centre not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserDTO"
          }
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the user"
          }
        }
      },
      "NewUser": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserDTO"
          }
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "The password of the user."
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "required": [
          "username",
          "firstName",
          "lastName",
          "phoneNumber",
          "email",
          "age",
          "userType",
          "centreId"
        ],
        "properties": {
          "username": {
            "type": "string",
            "description": "The username of the user."
          },
          "firstName": {
            "type": "string",
            "description": "The first name of the user."
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the user."
          },
          "email": {
            "type": "string",
            "description": "The email of the user."
          },
          "phoneNumber": {
            "type": "string",
            "description": "The phone number of the user."
          },
          "age": {
            "type": "integer",
            "description": "The age of the user."
          },
          "userType": {
            "type": "string",
            "enum": ["Vet", "Admin", "Volunteer"],
            "description": "The user type of the user."
          },
          "centreId": {
            "type": "string",
            "description": "The ID of the centre associated with the user."
          }
        }
      },
...
      "Scale": {
        "type": "object",
        "required": ["id", "scaleRef", "centreId", "isReserved"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the scale"
          },
          "scaleRef": {
            "type": "string",
            "description": "The refference of the scale"
          },
          "centreId": {
            "type": "string",
            "description": "ID of the centre the scale belongs to"
          },
          "isReserved": {
            "type": "boolean",
            "description": "Whether the scale is reserved or not"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}